Migrate ProductName to TS
