--- react/ProductName.js
@@ -1,3 +1,0 @@
-import ProductName from './components/ProductName/Wrapper'
-
-export default ProductName
