+++ react/package.json
@@ -39,6 +39,7 @@
     "@types/node": "^12.7.5",
     "@types/ramda": "^0.26.18",
     "@types/react": "^16.9.2",
+    "@types/react-content-loader": "3",
     "@types/react-dom": "^16.9.0",
     "@vtex/test-tools": "^3.1.0",
     "@vtex/tsconfig": "^0.3.0",
